Moved constants to domain file to trim up main icon.tsx file.
