+++ src/DNDTools/DiceRoller/Die.tsx
@@ -1,6 +1,8 @@
 import './die.css';
 import { rollDie } from './util';
-import { Icon, IconSize, IconType } from '../../SharedComponents/Icon/Icon';
+import { Icon } from '../../SharedComponents/Icon/Icon';
+import { IconSize } from '../../SharedComponents/Icon/domain';
+import { IconType } from '../../SharedComponents/Icon/domain';
 
 type Props = {
     diceType: number;
