+++ src/ErrorStates/ErrorBoundary.tsx
@@ -4,7 +4,8 @@ import { useEffect } from "react";
 import './route-error-boundary.css';
 import { Button } from "../SharedComponents/Button/Button";
 import { HomeIcon } from "../SharedComponents/Icon/Icons/HomeIcon";
-import { Icon, IconType } from "../SharedComponents/Icon/Icon";
+import { Icon } from "../SharedComponents/Icon/Icon";
+import { IconType } from '../SharedComponents/Icon/domain';
 
 export const RouterErrorBoundary = () => {
     const error = useRouteError();
