+++ src/HiddenPages/TerminalEditor/TerminalListItem.tsx
@@ -1,7 +1,8 @@
 
 import "./server-item.css";
 import { Button, ButtonTheme } from "../../SharedComponents/Button/Button";
-import { Icon, IconType } from "../../SharedComponents/Icon/Icon";
+import { Icon } from "../../SharedComponents/Icon/Icon";
+import { IconType } from '../../SharedComponents/Icon/domain';
 
 type Props = {
     label: string
