+++ src/HiddenPages/WeaponEffectEditor/WeaponEffectTag.tsx
@@ -1,7 +1,8 @@
 
 import './weapon-effect-tag.css';
 import { Button, ButtonTheme } from "../../SharedComponents/Button/Button";
-import { Icon, IconType } from '../../SharedComponents/Icon/Icon';
+import { Icon } from '../../SharedComponents/Icon/Icon';
+import { IconType } from '../../SharedComponents/Icon/domain';
 
 type Props = {
     tag: string;
