+++ src/RPGTools/DialogueTreeMaker/DialogueMaker/HiddenInfo/HiddenInfoItem.tsx
@@ -2,7 +2,8 @@
 import './hidden-info-item.css';
 import { Button, ButtonTheme } from "../../../../SharedComponents/Button/Button";
 import { HiddenInfo } from "../../domain/types";
-import { Icon, IconType } from '../../../../SharedComponents/Icon/Icon';
+import { Icon } from '../../../../SharedComponents/Icon/Icon';
+import { IconType } from '../../../../SharedComponents/Icon/domain';
 
 type Props = {
     hiddenInfo: HiddenInfo;
