+++ src/RPGTools/DialogueTreeMaker/DialogueMaker/HiddenInfo/UpdateHiddenInfoModal.tsx
@@ -6,7 +6,8 @@ import { Modal } from "../../../../SharedComponents/Modal/Modal";
 import { Button, ButtonTheme } from "../../../../SharedComponents/Button/Button";
 import { TextInput } from "../../../../SharedComponents/TextInput/TextInput";
 import { DescriptionCard } from "../../../../SharedComponents/DescriptionCard/DescriptionCard";
-import { Icon, IconType } from "../../../../SharedComponents/Icon/Icon";
+import { Icon } from "../../../../SharedComponents/Icon/Icon";
+import { IconType } from '../../../../SharedComponents/Icon/domain';
 
 type Props = {
     hiddenInfoToEdit: HiddenInfo | null;
