+++ src/RPGTools/DialogueTreeMaker/SkillTestMaker/SkillTestMaker.tsx
@@ -6,7 +6,8 @@ import { SkillTest, SkillTestDifficulty } from "../domain/types";
 import { Card } from "../../../SharedComponents/Card/Card";
 import { SkillTestDifficultyModal } from "./SkillTestDifficultyModal";
 import { Button, ButtonTheme } from "../../../SharedComponents/Button/Button";
-import { Icon, IconType } from "../../../SharedComponents/Icon/Icon";
+import { Icon } from "../../../SharedComponents/Icon/Icon";
+import { IconType } from '../../../SharedComponents/Icon/domain';
 
 type Props = {
     currentSkillTest: SkillTest;
