+++ src/SharedComponents/Icon/domain.ts
@@ -1,0 +1,32 @@
+export enum IconType {
+    CLOSE,
+    COPY,
+    DICE4,
+    DICE6,
+    DICE8,
+    DICE10,
+    DICE12,
+    DICE20,
+    DOANLOAD,
+    DRAG,
+    EDIT_FILE,
+    GEARS,
+    HOME,
+    PENCIL,
+    PLUS,
+    TAUNT,
+    TRASH,
+    UPLOAD_FILE,
+    VERY_DISSATISFIED
+};
+
+export enum IconTheme { LIGHT, DARK, GREEN };
+
+export enum IconSize {
+    SMALL = "24px",
+    MEDIUM = "32px",
+    LARGE = "40px",
+    EXTRA_LARGE = "80px",
+    RESPONSIVE = "70%"
+};
+
